# [core] perf improvements for Assembly attributes reflection