--- a/Xamarin.Forms.Core/ReflectionExtensions.cs
+++ b/Xamarin.Forms.Core/ReflectionExtensions.cs
@@ -47,13 +47,12 @@ namespace Xamarin.Forms.Internals
 
 		internal static object[] GetCustomAttributesSafe(this Assembly assembly,  Type attrType)
 		{
-			object[] attributes = null;
 			try
 			{
 #if NETSTANDARD2_0
-				attributes = assembly.GetCustomAttributes(attrType, true);
+				return assembly.GetCustomAttributes(attrType, true);
 #else
-				attributes = assembly.GetCustomAttributes(attrType).ToArray();
+				return assembly.GetCustomAttributes(attrType).ToArray();
 #endif
 			}
 			catch (System.IO.FileNotFoundException)
@@ -62,7 +61,7 @@ namespace Xamarin.Forms.Internals
 				Log.Warning(nameof(Registrar), "Could not load assembly: {0} for Attribute {1} | Some renderers may not be loaded", assembly.FullName, attrType.FullName);
 			}
 
-			return attributes;
+			return null;
 		}
 
 		public static Type[] GetExportedTypes(this Assembly assembly)